requirements.txt
numpy
pandas
ibllib
ONE-api
torch
behavior_models @ git+https://github.com/int-brain-lab/behavior_models.git